Deployment/requirements.txt
streamlit
tensorflow==2.15.1
numpy==1.26.4
Pillow==11.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7